Update user.js: I added the automatic password encription and enum UserType
I created a method that is automatically called when the new account is created in a way that the password is authomatically ashed with the bcrypt library.
I fixed a problem with the userType and I set it as enum for simplicity. 
I removed the automatic date set to "now" for the date of birthday.